The application is now capable of drawing a plan from a preset file. Zooming and panning of the plan is implimented too.

+++ GBPDMA/src/com/gbpdma/logic/LocationPoint.java
@@ -4,8 +4,8 @@ public class LocationPoint {
 
 	private double longitude;
 	private double latitude;
-	private double x;
-	private double y;
+	private float x;
+	private float y;
 	
 	public LocationPoint(double longitude, double latitude)
 	{
@@ -25,16 +25,16 @@ public class LocationPoint {
 	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
-	public double getX() {
+	public float getX() {
 		return x;
 	}
-	public void setX(double x) {
+	public void setX(float x) {
 		this.x = x;
 	}
-	public double getY() {
+	public float getY() {
 		return y;
 	}
-	public void setY(double y) {
+	public void setY(float y) {
 		this.y = y;
 	}
